requirements.txt
opencv-python
pillow
pytesseract
imutils
numpy
datetime
requests
